refactor: Clean up result building
Result building and printing was tedious to maintain given that the
functions ended up requiring fields rather than the entire record. This
made some functions' params/args to be ridiculously lengthy.

Since the Result record has fields similar to Color, it was better to
remove the individual `hexString`, `rgb`, etc fields and just use
Color as a field. This also removes the need for the IdMap as the
terminal IDs can be extracted from the Colors granted they exist.