feat: Add option for merged results
This commit adds a `--merge` flag to merge the results from different
comparison sources (i.e input for `--xterm256`, `--file`, etc).
Originally, if multiple comparison sources are given, the results are
isolated to themselves.

For example:
$ ./<program> "<hex string>" --xterm256 --file <path/to/json>

Compared to: xterm256
...
...

Compared to: <path/to/json>
...
...

The new `--merged` flag allows for the results to be merged to together
for a single result output.

For Example:
$ ./<program> "<hex string>" --xterm256 --file <path/to/json> --merge

Compared to: merged
...
...
...